create user auth context for all app + permissions handler
