+++ src/app/games/layout.tsx
@@ -1,12 +1,17 @@
+'use client'
+
 import React from "react";
 import { links } from "@/utils/links";
 import Navbar from "@/components/Navbar";
+import { withPageAuthorized } from "@/components/AuthValidation";
 
-export default function layout({ children }: { children: React.ReactNode }) {
+export default withPageAuthorized(function layout({ children }: { children: React.ReactNode }) {
   return (
     <div className='mb-44'>
       <Navbar text={links.games} />
       {children}
     </div>
   );
-}
+},{
+  returnTo: "/games"
+})
