+++ src/components/Menu/SideMenu.tsx
@@ -1,8 +1,11 @@
+'use client'
+
 import React from "react";
 import { deleteLocalStorageUserWithJwt } from "@/services/auth/localStorageService";
-import { links } from "@/utils/links";
+import { links, linksPermissions } from "@/utils/links";
 import { useRouter } from "next/navigation";
 import { Plus, Docs, Controller, Joystick, Question, Hamburger, Logout } from "../Icons";
+import { useCustomUserContext } from "@/app/context/userStore";
 
 const links_svgs = {
   create: <Plus />,
@@ -15,12 +18,18 @@ const links_svgs = {
 
 export function SideMenu({
   text,
-  hideMenu,
+  hideMenu
 }: {
   text: string;
   hideMenu: () => void;
 }) {
+  const {user} = useCustomUserContext();
   const router = useRouter();
+
+  if(!user) {
+    return null;
+  }
+  console.log(user);
 
   return (
     <aside
@@ -33,7 +42,7 @@ export function SideMenu({
         </button>
       </div>
       <nav className="flex flex-col h-full my-10">
-        {Object.keys(links).map((link) => (
+        {Object.keys(links).filter((link) => user?.permissions?.includes(linksPermissions[link as keyof typeof linksPermissions])).map((link) => (
           <a
             href={link}
             key={link}
