+++ src/hooks/useApiQuery.ts
@@ -1,12 +1,9 @@
-import { useRouter } from 'next/navigation';
+import { redirect, useRouter } from 'next/navigation';
 import { useEffect } from 'react';
 import { useQuery, useQueryClient } from '@tanstack/react-query';
 import Swal from 'sweetalert2';
 import { ApiResponse } from '../services/api';
-import {
-  deleteLocalStorageUserWithJwt,
-  getLocalStorageUserWithJwt,
-} from '../services/auth/localStorageService';
+import { useCustomUserContext } from '@/app/context/userStore';
 
 type apiQueryLoading = {
   data: undefined;
@@ -50,27 +47,30 @@ const error_response: apiQueryError = {
 
 export default function useApiQuery<T>(
   query_key: string,
-  queryFn: (query_key: string, access_token?: string) => Promise<ApiResponse<T>>,
-  access_token?: string
+  queryFn: (query_key: string, access_token?: string) => Promise<ApiResponse<T>>
 ): apiQueryResponse<T> {
   const router = useRouter();
   const query_client = useQueryClient();
+  const { accessToken } = useCustomUserContext();
   const {
     data,
     isLoading: is_loading,
     fetchStatus: fetch_status,
-  } = useQuery([query_key], () => queryFn(query_key, access_token), {
+  } = useQuery([query_key, accessToken], () => queryFn(query_key, accessToken ?? undefined), {
     staleTime: Infinity,
     cacheTime: Infinity,
   });
 
   useEffect(() => {
-    console.log(`fetched key ${query_key}`);
-    console.log(data)
+    console.debug(`fetched key ${query_key}`);
+    console.debug(data)
 
     // * If api failed, we inform the user
     // * we need to check fetch_status, otherwise the error may be caused by an old cached error
     if (data && data.status !== 'success' && fetch_status !== 'fetching') {
+      if (data.code === 401 ) {
+        return redirect("/api/auth/login?returnTo=" + (window.location.href ?? "/"));
+      }
       Swal.fire({
         title: "Qualcosa e' andato storto",
         html: data.message,
@@ -78,17 +78,6 @@ export default function useApiQuery<T>(
         allowOutsideClick: false,
       }).then(() => {
         query_client.invalidateQueries([query_key]);
-
-        // * unauthorized, the user does not a (valid) token
-        if (data.code === 401) {
-          // * unauthorized, if the token exists it's invalid
-          const token = getLocalStorageUserWithJwt().access_token;
-          if (token !== null) {
-            deleteLocalStorageUserWithJwt();
-          }
-          // * redirect to login page
-          router.push('login');
-        }
       });
     }
   }, [data, router, query_client, query_key, fetch_status]);
