+++ src/utils/links.ts
@@ -6,3 +6,12 @@ export const links = {
   tutorial: "Come funziona",
   vocabularies: "Gestisci vocabolari"
 } as const;
+
+export const linksPermissions = {
+  create: "create:rules",
+  rules: "read:rules",
+  elements: "read:elements",
+  games: "delete:games",
+  tutorial: "read:tutorial",
+  vocabularies: "update:vocabularies"
+} as const
